requirements.txt
Flask
gunicorn
requests
aiohttp
protobuf
pycryptodome
redis
